fast MPCC working
The implementation is working as intended however it is not running fast in speed. See what you need to do to make it go faster ie optimised line
The parameter change does not seem to effect the racing line that it takes, the change in lag and contour should make the tracking better or worse and it is not doing that.